requirements.txt
langchain>=0.1.0
openai>=1.10.0
chromadb>=0.4.22
pypdf>=3.15.0
streamlit>=1.30.0
python-dotenv>=1.0.0